# Conflicts:
#	Assets/02_Scripts/Monster/MonsterObj/AutoTurret.cs
#	Assets/02_Scripts/Monster/MonsterObj/NebulaVanguard.cs
#	Assets/02_Scripts/Skill/Skill_Four.cs

+++ Assets/02_Scripts/Skill/ReinforceSkill/SPAttackReinforce.cs
@@ -69,7 +69,7 @@ public class Skill_Two_SPAttack : ActiveTech
 
     public override event Action OnTechUsed;
 
-    private float originalSPAttack;
+    private int originalSPAttack;
 
 
     /// <summary>
